# [libc++] Optimize ranges::minmax